unbundle react generator
