+++ packages/core/src/generators/react/index.ts
@@ -1,1 +1,2 @@
 export * from './generator';
+export * from './types';
