+++ packages/core/src/generators/react/types.ts
@@ -1,0 +1,11 @@
+import { BaseTranspilerOptions } from '../../types/transpiler';
+
+export interface ToReactOptions extends BaseTranspilerOptions {
+  stylesType?: 'emotion' | 'styled-components' | 'styled-jsx' | 'react-native';
+  stateType?: 'useState' | 'mobx' | 'valtio' | 'solid' | 'builder';
+  format?: 'lite' | 'safe';
+  type?: 'dom' | 'native';
+  preact?: boolean;
+  forwardRef?: string;
+  experimental?: any;
+}
